+++ src/lips.js
@@ -6306,17 +6306,38 @@
                             pair.car.cdr instanceof Pair &&
                             pair.car.cdr.car instanceof Pair &&
                             LSymbol.is(pair.car.cdr.car.car, 'unquote-splicing')) {
+                            const rest = pair.car.cdr;
                             return new Pair(
-                                new LSymbol('unquote'),
-                                unquote_splice(pair.car.cdr, unquote_cnt + 2, max_unq)
+                                new Pair(
+                                    new LSymbol('unquote'),
+                                    unquote_splice(rest, unquote_cnt + 2, max_unq)
+                                ),
+                                nil
                             );
                         } else if (pair.car.cdr instanceof Pair &&
-                                   pair.car.cdr.cdr !== nil &&
-                                   !(pair.car.cdr.car instanceof Pair)) {
-                            // same as in guile if (unquote 1 2 3) it should be
-                            // spliced - scheme spec say it's unspecify but it
-                            // work like in CL
-                            return pair.car.cdr;
+                                   pair.car.cdr.cdr !== nil) {
+                            if (pair.car.cdr.car instanceof Pair) {
+                                // values inside unquote are lists
+                                const result = [];
+                                return (function recur(node) {
+                                    if (node === nil) {
+                                        return Pair.fromArray(result);
+                                    }
+                                    return unpromise(evaluate(node.car, {
+                                        env: self,
+                                        dynamic_scope,
+                                        error
+                                    }), function(next) {
+                                        result.push(next);
+                                        return recur(node.cdr);
+                                    });
+                                })(pair.car.cdr);
+                            } else {
+                                // same as in guile if (unquote 1 2 3) it should be
+                                // spliced - scheme spec say it's unspecify but it
+                                // work like in CL
+                                return pair.car.cdr;
+                            }
                         }
                     }
                     if (LSymbol.is(pair.car, 'quote')) {
@@ -6340,18 +6361,22 @@
                         if (pair.cdr instanceof Pair) {
                             if (pair.cdr.cdr !== nil) {
                                 if (pair.cdr.car instanceof Pair) {
-                                    let list = nil;
+                                    // TODO: test if this part is needed
+                                    // this part was duplicated in previous section
+                                    // if (LSymbol.is(pair.car.car, 'unquote')) {
+                                    // so this probably can be removed
+                                    const result = [];
                                     // evaluate all values in unquote
                                     return (function recur(node) {
                                         if (node === nil) {
-                                            return list;
+                                            return Pair.fromArray(result);
                                         }
                                         return unpromise(evaluate(node.car, {
                                             env: self,
                                             dynamic_scope,
                                             error
                                         }), function(next) {
-                                            list = new Pair(next, list);
+                                            result.push(next);
                                             return recur(node.cdr);
                                         });
                                     })(pair.cdr);